Added rupiah formatting

+++ BudgetingMethods/FrugalMethod.java
@@ -3,11 +3,11 @@ import BudgetingPlan.Budget;
 
 public class FrugalMethod extends Method {
     public Budget calculateBudget(long income){
-        int save =(int) Math.round(income * 0.05);
-        int invest =(int) Math.round(income * 0.6);
+        long save = Math.round(income * 0.05);
+        long invest = Math.round(income * 0.6);
 
-        int wants = (int) Math.round(income * 0.1);
-        int needs = (int) Math.round(income * 0.25);
+        long wants = Math.round(income * 0.1);
+        long needs = Math.round(income * 0.25);
 
         return new Budget(save, invest, wants, needs);
     }
